// src/types/Stripe.ts
// Generated by https://quicktype.io

export interface CheckoutResponse {
    ref:                                   any;
    after_expiration:                     null;
    allow_promotion_codes:                null;
    amount_subtotal:                      number;
    amount_total:                         number;
    automatic_tax:                        AutomaticTax;
    billing_address_collection:           null;
    cancel_url:                           string;
    client_reference_id:                  string;
    client_secret:                        null;
    consent:                              null;
    consent_collection:                   null;
    created:                              number;
    currency:                             string;
    currency_conversion:                  null;
    custom_fields:                        any[];
    custom_text:                          CustomText;
    customer:                             string;
    customer_creation:                    null;
    customer_details:                     CustomerDetails;
    customer_email:                       null;
    expires_at:                           number;
    id:                                   string;
    invoice:                              null;
    invoice_creation:                     null;
    livemode:                             boolean;
    locale:                               string;
    metadata:                             Metadata;
    mode:                                 string;
    object:                               string;
    payment_intent:                       null;
    payment_link:                         null;
    payment_method_collection:            string;
    payment_method_configuration_details: PaymentMethodConfigurationDetails;
    payment_method_options:               null;
    payment_method_types:                 string[];
    payment_status:                       string;
    phone_number_collection:              PhoneNumberCollection;
    recovered_from:                       null;
    setup_intent:                         null;
    shipping_address_collection:          null;
    shipping_cost:                        null;
    shipping_details:                     null;
    shipping_options:                     any[];
    status:                               string;
    submit_type:                          null;
    subscription:                         null;
    success_url:                          string;
    total_details:                        TotalDetails;
    ui_mode:                              string;
    url:                                  string;
}

export interface AutomaticTax {
    enabled: boolean;
    status:  null;
}

export interface CustomText {
    shipping_address:            null;
    submit:                      null;
    terms_of_service_acceptance: null;
}

export interface CustomerDetails {
    address:    null;
    email:      string;
    name:       null;
    phone:      null;
    tax_exempt: string;
    tax_ids:    null;
}

export interface Metadata {
    leadType:      string;
    leadsEstimate: string;
    marketShare:   string;
    src:           string;
    uid:           string;
    zipCode:       string;
}

export interface PaymentMethodConfigurationDetails {
    id:     string;
    parent: null;
}

export interface PhoneNumberCollection {
    enabled: boolean;
}

export interface TotalDetails {
    amount_discount: number;
    amount_shipping: number;
    amount_tax:      number;
}