// src/types/UserPublicData.ts
export interface UserPublicData {
    avatar?: string;
    bio?: string;
    broker?: string;
    email?: string;
    facebook?: string;
    instagram?: string;
    license?: string;
    linkedin?: string;
    name?: string;
    phone?: string;
    twitter?: string;
    website?: string;
    youtube?: string;
    youtubeVideo?: string;
}